fooditems on homescreen

+++ app/src/main/java/com/shaun/foodnut/models/foodparser/Measure.kt
@@ -4,5 +4,5 @@ data class Measure(
     val label: String,
     val qualified: List<Qualified>,
     val uri: String,
-    val weight: Int
+    val weight: Double
 )
